fix: 모달 깜박거리는 현상 제거
